[MID-94] fully functoinal coordinates

+++ prisma/migrations/20240321074518_added_address_location/migration.sql
@@ -6,5 +6,5 @@
 
 */
 -- AlterTable
-ALTER TABLE "address" ADD COLUMN     "latitude" REAL NOT NULL,
-ADD COLUMN     "longitude" REAL NOT NULL;
+ALTER TABLE "address" ADD COLUMN     "latitude" DOUBLE PRECISION NOT NULL,
+ADD COLUMN     "longitude" DOUBLE PRECISION NOT NULL;
